Add files via upload
Recommendation system script runs two different prediction models. One is a statistical prediction approach that aims to utilize item and user biases to create a prediction rating on a scale from 1 - 5, this method leverages mean operations to derive a prediction value. On the other hand, the more complex prediction process is one based on a gradient descent approach, in this model the algorithm attempts to minimize an objective function for item and user biases, again these biases are used to generate a prediction.